test: run fsmeta contract on raftstore
